requirements.txt
duckdb
pandas
tabulate
requests
yato-lib
dlt
plotly
kaleido